refactor!(common): #388 moved WriteOscalModel to the oscal package (complete-schema)
