+++ src/pkg/common/oscal/complete-schema.go
@@ -1,7 +1,15 @@
 package oscal
 
 import (
+	"bytes"
+	"encoding/json"
+	"os"
+	"path/filepath"
+
+	"github.com/defenseunicorns/go-oscal/src/pkg/utils"
 	oscalTypes_1_1_2 "github.com/defenseunicorns/go-oscal/src/types/oscal-1-1-2"
+	"github.com/defenseunicorns/lula/src/pkg/message"
+	yamlV3 "gopkg.in/yaml.v3"
 	"sigs.k8s.io/yaml"
 )
 
@@ -12,6 +20,60 @@ func NewOscalModel(data []byte) (*oscalTypes_1_1_2.OscalModels, error) {
 		return nil, err
 	}
 	return &oscalModel, nil
+}
+
+// WriteOscalModel takes a path and writes content to a file while performing checks for existing content
+// supports both json and yaml
+func WriteOscalModel(filePath string, model *oscalTypes_1_1_2.OscalModels) error {
+
+	// if no path or directory add default filename
+	if filepath.Ext(filePath) == "" {
+		filePath = filepath.Join(filePath, "oscal.yaml")
+	}
+
+	if err := utils.IsJsonOrYaml(filePath); err != nil {
+		return err
+	}
+
+	if _, err := os.Stat(filePath); err == nil {
+		// If the file exists - read the data into the model
+		existingFileBytes, err := os.ReadFile(filePath)
+		if err != nil {
+			return err
+		}
+		existingModel, err := NewOscalModel(existingFileBytes)
+		if err != nil {
+			return err
+		}
+		// Merge the existing model with the new model
+		// re-assign to perform common operations below
+		model, err = MergeOscalModels(existingModel, model)
+		if err != nil {
+			return err
+		}
+	}
+
+	var b bytes.Buffer
+
+	if filepath.Ext(filePath) == ".json" {
+		jsonEncoder := json.NewEncoder(&b)
+		jsonEncoder.SetIndent("", "  ")
+		jsonEncoder.Encode(model)
+	} else {
+		yamlEncoder := yamlV3.NewEncoder(&b)
+		yamlEncoder.SetIndent(2)
+		yamlEncoder.Encode(model)
+	}
+
+	err := utils.WriteOutput(b.Bytes(), filePath)
+	if err != nil {
+		return err
+	}
+
+	message.Infof("OSCAL artifact written to: %s", filePath)
+
+	return nil
+
 }
 
 func MergeOscalModels(existingModel *oscalTypes_1_1_2.OscalModels, newModel *oscalTypes_1_1_2.OscalModels) (*oscalTypes_1_1_2.OscalModels, error) {
